Tidy registerWithRequire code
